% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.OmicsLens}
\alias{print.OmicsLens}
\title{Print an OmicsLens object}
\usage{
\method{print}{OmicsLens}(x, ...)
}
\arguments{
\item{x}{An OmicsLens object.}

\item{...}{Ignored.}
}
\description{
Print an OmicsLens object
}
