% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary.OmicsLens}
\alias{summary.OmicsLens}
\title{Summarise an OmicsLens object}
\usage{
\method{summary}{OmicsLens}(object, ...)
}
\arguments{
\item{object}{An OmicsLens object.}

\item{...}{Ignored.}
}
\description{
Summarise an OmicsLens object
}
